Modify embedding workload input format.
Instead of specifying query_directory or query_order in the benchmark's YAML,
we pass in workload-name as a CLI argument.

Assuming that the gym is generating the workload, (benchmark, workload-name)
is enough to identify the workload's folder and its corresponding order.txt.
The format of order.txt is currently:
QueryId,AbsPathToQuerySQL[,Weight]