Replicate TPC-H SF10 tuning results (#28)
**Summary**: Can now replicate TPC-H SF10 **tuning** results on dev 4.
Note that **embedding** has not yet been replicated.

**Demo**:
After running HPO on dev 4 for 18 hours, we get to a runtime of 110s. We
run with `--duration=4`, and 110s matches the performance in the Proto-X
paper after 4 hours of tuning.
![Screenshot 2024-04-05 at 13 12
03](https://github.com/cmu-db/dbgym/assets/20631215/16babc69-4b42-434d-b1b3-eede0010f1da)

**Details**:
* Because I was only replicating tuning, I used an already-trained
embedder. To replicate, use the `dbgym_manual/embedding/` folder on
dev4.
* The command I used was `python3 task.py --no-startup-check tune protox
agent hpo tpch --scale-factor 10 --max-concurrent 4 --duration 4
--intended-pgdata-hardware ssd --pgdata-parent-dpath
/mnt/nvme1n1/phw2/dbgym_tmp/`