Replay Proto-X (#32)
**Summary**: Can now replay a full tuning run from Proto-X. This is used
to see how each step of tuning would have done without query timeouts
and without Boot enabled.

**Demo**:
The image shows the data of a replayed run of TPC-H SF0.01 without Boot
enabled during tuning. For each step of tuning, the replay shows the #
of queries executed during the original run (which may be < 22 if the
workload timed out), the # of queries that timed out during the original
run, and whether the workload timed out. It also shows this same
information about the replay. You can see that the replayed times are
always >= the original times, which makes sense. Whenever the original
run is "22,0,False" (i.e. 22 executed, 0 timed out, workload didn't time
out), the replayed time matches closely.
![Screenshot 2024-04-27 at 14 49
15](https://github.com/cmu-db/dbgym/assets/20631215/dbf87c34-4340-4d46-b91b-61f3fefa69ef)

**Details**:
* Migrated the pipeline which pickles actions (DBMS configuration
changes) while tuning and replays actions during replay to get the DBMS
into the correct state.
* Related to the above, modified the pipeline to store both the best
per-query knobs found during `execute_variations()` as well as all
per-query knob variations tried. We can either replay the best variation
or all variations. This is especially useful if the workload timed out
in the original run, in which case the "best" variation is a misnomer as
it is simply an arbitrary variation.
* Made Proto-X log additional information while tuning about the # of
executed queries, # of timed out queries, and whether the workload timed
out. During replay, all this information may now be utilized.
* Fixed a bug where `reset()` was overwriting the logged replay
information for a step, leading to a mismatch between the dumped
`action.pkl` file (which contains the DBMS configuration state) and the
`run.raw.csv` file (which contains the runtime information of the
workload during that state).
* Refactored all symlinks to have the `.link` extension to fix a subtle
bug where a replay would overwrite the `output.log` file of the original
run.
* Standardized whether the page cache is dumped during tuning and replay
(it's now not dumped in either case).
* Made CLI options for the time to run the RL agent and whether Boot is
enabled more fine-grained such that these values can differ between HPO
and tune.