Integration test CI (#37)
**Summary**: Added a CI for running integration tests. Fixed a few bugs
to get it working.

**Demo**:
[Successful
CI](https://github.com/cmu-db/dbgym/actions/runs/10242135761/job/28331697565)
![Screenshot 2024-08-04 at 21 08
29](https://github.com/user-attachments/assets/f70cbc54-4cf7-473a-a37c-4546e60b63ee)

**Details**:
* Made the runner use the `runner` user instead of `root` so we can
start Postgres in it.
* Made new DBGym config for integration test so that the ports and
workspace don't conflict (yay config file!). This is not currently
necessary since we run it in a Docker container but it's a good
precaution.
* In the integration test, we run the full chain (with replay) twice so
that we test it both without and with cached results in the workspace.
We make sure to delete the workspace before running twice (again not
necessary because we're using Docker but is a good precaution).