Postgres start/stop demo (#45)
**Summary**: Basic demo letting you start and stop Postgres.

**Demo**:
(Yes, this is a demo of the demo). The demo uses `pgrep postgres` to
show that no Postgres instance is running. Then the user is able to
start/stop Postgres on the web UI. The results of that are verified
using `pgrep postgres`.


https://github.com/user-attachments/assets/e407df5d-62b6-4c14-90c8-7602d2d8dce1

**Details**:
* Uses `streamlit` for the frontend.
* Uses the recently factored out `PostgresConn` in the backend.
* Right now is only deployable locally. In the future we'll want to
deploy it on a CMU-DB machine so people can play around with it.
* It currently relies on the same workspace used in development. In the
future, we may want to give the demo its own workspace. The setup
process of this workspace would then require more scripting.