Query execution interface (#48)
**Summary**: Added query execution interface in `env/`.

**Demo**:
A passing unit test.
<img width="1092" alt="Screenshot 2024-12-21 at 12 38 33"
src="https://github.com/user-attachments/assets/1556d975-f5d3-4a2e-b571-00288fa2f355"
/>

**Details**:
* The specific interface is the `time_query()` function in
`PostgresConn`.
* I chose to have a **per-query interface** instead of a per-workload
interface because different agents might have different ways of
executing the queries of an OLAP workload. For example, Proto-X executes
multiple "variations" of each query when executing a workload. A future
agent might also only execute a subset of queries in a workload.
* The interface allows you to specify per-query knobs and a timeout. It
returns the **runtime but not the results** of the query. The reason it
does not return the results is so that it's **compatible with Boot**
(which estimates the runtime but not the results).