Workspace refactor (#66)
**Summary**: Refactored the workspace for the new submodule
architecture.

**Demo**:
Added 22 tests in a new file (unittest_workspace.py) that are passing.
<img width="835" alt="Screenshot 2024-12-28 at 16 12 35"
src="https://github.com/user-attachments/assets/44c23907-de30-441d-8229-4c332002a8de"
/>

**Details**:
* The refactor is done but is **not integrated**. The old functions are
all there. New versions of `save_file()` and `link_result()` are in
`DBGymWorkspace` but are not used anywhere outside of the unit tests.
* I got rid of the `[codebase]` component in `task_runs/` and
`symlinks/` and replaced it with the `[agent]` component. This makes it
possible to **know where files from other agents are stored**, which is
crucial in the new submodule architecture.
* In the monolith architecture, we knew the codebase structure of agents
so we could build the `[codebase]` accordingly. Now, we don't know this
so we need to decouple communication from the codebase structure of the
agent.
* Note that only `symlinks/` uses`[agent]`. `task_runs/` doesn't use
either `[codebase]` or `[agent]` anymore for further simplicity.
* I got rid of the `[org]` component as well for simplicity. The user
can add their own organization if they like.
* I renamed `DBGymConfig` to `DBGymWorkspace` to better reflect its
function. I changed its initializer to only take in the workspace path
instead of the config path as well.
* I moved all the filesystem unittest helpers to their own file since
both `unittest_workspace.py` and `unittest_clean.py` use them. This is
more than just an aesthetic change. If we didn't move them and instead
had `unittest_workspace.py` import `unittest_clean.py`, we would run the
tests in `unittest_clean.py` twice.